Refactoring of namings
