+++ cmd/deepchain/cmd/genaccounts.go
