+++ docker-compose.yml
@@ -9,9 +9,9 @@ services:
       - "1317:1317"
     environment:
       - ID=0
-      - LOG=${LOG:-cyber.log}
+      - LOG=${LOG:-deepchain.log}
     volumes:
-      - ./build:/cyber:Z
+      - ./build:/deepchain:Z
     networks:
       localnet:
         ipv4_address: 192.168.10.2
@@ -23,9 +23,9 @@ services:
       - "26659-26660:26656-26657"
     environment:
       - ID=1
-      - LOG=${LOG:-cyber.log}
+      - LOG=${LOG:-deepchain.log}
     volumes:
-      - ./build:/cyber:Z
+      - ./build:/deepchain:Z
     networks:
       localnet:
         ipv4_address: 192.168.10.3
@@ -35,11 +35,11 @@ services:
     image: "cybercongress/cyber"
     environment:
       - ID=2
-      - LOG=${LOG:-cyber.log}
+      - LOG=${LOG:-deepchain.log}
     ports:
       - "26661-26662:26656-26657"
     volumes:
-      - ./build:/cyber:Z
+      - ./build:/deepchain:Z
     networks:
       localnet:
         ipv4_address: 192.168.10.4
@@ -49,11 +49,11 @@ services:
     image: "cybercongress/cyber"
     environment:
       - ID=3
-      - LOG=${LOG:-cyber.log}
+      - LOG=${LOG:-deepchain.log}
     ports:
       - "26663-26664:26656-26657"
     volumes:
-      - ./build:/cyber:Z
+      - ./build:/deepchain:Z
     networks:
       localnet:
         ipv4_address: 192.168.10.5
