+++ networks/local/Makefile
@@ -1,4 +1,4 @@
-# Makefile for the "cyber" docker image.
+# Makefile for the "deepchain" docker image.
 
 all:
 	docker build --tag cybercongress/cyber cybernode
